moved resolvers under Resolver namespace
